+++ sopa/segmentation/methods/__init__.py
@@ -3,3 +3,4 @@ from ._comseg import comseg
 from ._dummy import dummy_method
 from ._baysor import baysor
 from ._custom import custom_staining_based
+from ._stardist import stardist